+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 ## [Unreleased]
 
-- Support `valgrind` and `wasm-pack`.
+- Support `valgrind`, `wasm-pack`, and `wasmtime`.
 
 ## [1.0.3] - 2022-01-06